Initial Development Release
This is the initial development release.  It has a whole bunch of the
ArduPlane code removed and a tiny bit of new code added, mainly in the
attitude control routine.  This build will compile without errors but is
totally untested.  Use only as a basis for further development, not
testing.